requirements.txt
requests
flask
folium
pywebview
tk